fetch price fix & prepare address book features
